gpu: nvgpu: cond: use u32 for COND_WAIT timeout
The type for the timeout parameter to the NVGPU_COND_WAIT and
NVGPU_COND_WAIT_INTERRUPTIBLE macros was too weak. This updates these
macros to require a u32 for the timeout.

Users of the macros are updated to be compliant as necessary.

This addresses MISRA 10.3 violations for implicit conversions of types
of different size or essential type.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2017503
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/fifo/channel.c
@@ -286,7 +286,7 @@ void gk20a_wait_until_counter_is_N(
 		if (NVGPU_COND_WAIT(
 			    c,
 			    nvgpu_atomic_read(counter) == wait_value,
-			    5000) == 0) {
+			    5000U) == 0) {
 			break;
 		}
 
@@ -1832,7 +1832,7 @@ static int gk20a_channel_poll_worker(void *arg)
 {
 	struct gk20a *g = (struct gk20a *)arg;
 	struct gk20a_worker *worker = &g->channel_worker;
-	unsigned long watchdog_interval = 100; /* milliseconds */
+	u32 watchdog_interval = 100; /* milliseconds */
 	struct nvgpu_timeout timeout;
 	int get = 0;
